Update MRZ extraction to return the longest MRZ fragment
Also allow white spaces between symbols when searching surname and name
separator.

Also add a test when two patterns are provided in the OCR texts, but the
first one contains \n in the beginning of the MRZ, hence is not
recognized by our startegy and another pattern has to be used.